Update MainActivity.java
Removed fab code from this and mainActivity.xml, because the way i have main activity structured it is purely a container and the task list is the main fragment that starts in the app and then it switches like the band database when the fab is clicked. Since it is similar to the band database, I think the fab code belongs in the task list fragment file just like how the onclicklistener is set up in the band list fragment for the band app.